Bug fix for running with `paired=FALSE`.
The assumption that the first value column on the `cov` object is a score column
does not hold when running fragCounter with `paired=FALSE` - see the logic in `PrepareCov`
where a different bamUtils function is used. The `count` column is set in that branch
though and is the correct column to use here.

Preserving the initial logic if `count` column does not exist though in case other
usages depend on that behavior.